+++ Core.lua
@@ -12,7 +12,7 @@
 local ADDON, ns = ...
 local Core = WickCore
 
-ns.version = "0.1.0"
+ns.version = "0.2.0"
 
 local PROFILE_DEFAULTS = {
     window   = {},